# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Q-Wash API
  description: >
    Car-washing queue API: phone + SMS-OTP auth, washing point / service
    CRUD, capacity-aware booking availability, a live queue board, and
    booking history. See `docs/API.md` in the repo for the full prose
    contract and `docs/DATA_MODEL.md` for the schema and algorithms this
    spec is generated alongside.
  version: "1.0.0"
servers:
  - url: /
    description: Same-origin (this server also hosts the docs page)
...
security: []
paths:
...
  /api/v1/washing-points/{id}/services:
    parameters:
      - $ref: "#/components/parameters/washingPointId"
    get:
      tags: [Services]
      operationId: listServices
      summary: List services at a washing point
      description: Public. Each service includes its nested `price_options`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceList" }
    post:
      tags: [Services]
      operationId: createService
      summary: Create a service
      description: >
        Requires role `staff` or `admin`, scoped to their own washing point
        for staff (404 for another point, same as a nonexistent id).
        `price_options` must be non-empty; if none is marked `is_default`,
        the first is auto-promoted; more than one marked default is 400
        `multiple_default_price_options`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/services/{id}:
    parameters:
      - $ref: "#/components/parameters/serviceId"
    get:
      tags: [Services]
      operationId: getService
      summary: Get a service
      description: Public.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Services]
      operationId: updateService
      summary: Partially update a service
      description: >
        Requires role `staff` or `admin`, scoped to their own washing point
        for staff. Updates the service's own fields only — not its price
        options.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Services]
      operationId: deactivateService
      summary: Deactivate a service
      description: >
        Requires role `staff` or `admin`, scoped to their own washing point
        for staff. Sets `is_active=false`.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deactivated }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/services/{id}/price-options:
    parameters:
      - $ref: "#/components/parameters/serviceId"
    post:
      tags: [Services]
      operationId: addPriceOption
      summary: Add a price option to a service
      description: >
        Requires role `staff` or `admin`, scoped to their own washing point
        for staff. If `is_default: true`, any existing default for that
        service is unset first.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PriceOptionCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceOption" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/price-options/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Services]
      operationId: updatePriceOption
      summary: Partially update a price option
      description: >
        Requires role `staff` or `admin`, scoped to the price option's
        parent service's washing point for staff (resolved via an extra
        lookup, since the id alone doesn't reveal it). Explicitly setting
        `is_default: false` on the current sole default is rejected — 400
        `cannot_unset_default`; mark another one default instead.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PriceOptionUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceOption" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Services]
      operationId: deletePriceOption
      summary: Remove a price option
      description: >
        Requires role `staff` or `admin`, scoped to the price option's
        parent service's washing point for staff (same extra-lookup note
        as PATCH). 409 `last_price_option` if it's the service's only
        option; 409 `price_option_in_use` if any booking references it. If
        the removed option was the default, another remaining one is
        auto-promoted.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from OTP verify/refresh, e.g. `Authorization: Bearer <token>`"
  parameters:
    washingPointId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    serviceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed, or invalid/expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or hidden as not-found to avoid leaking existence)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict (e.g. slot no longer available, invalid transition)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_name }
            message: { type: string, example: "name is required (max 255 chars)" }
          required: [code, message]
      required: [error]
...
    PriceOption:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        price_cents: { type: integer, minimum: 0 }
        is_default: { type: boolean }
      required: [id, name, price_cents, is_default]
    PriceOptionInput:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        price_cents: { type: integer, minimum: 0 }
        is_default: { type: boolean, default: false }
      required: [name, price_cents]
    PriceOptionCreate:
      allOf:
        - $ref: "#/components/schemas/PriceOptionInput"
    PriceOptionUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        price_cents: { type: integer, minimum: 0 }
        is_default:
          type: boolean
          description: Explicitly setting false on the current sole default is rejected (400 cannot_unset_default).
    Service:
      type: object
      properties:
        id: { type: string, format: uuid }
        washing_point_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        duration_minutes: { type: integer, minimum: 1 }
        picture_url: { type: string, nullable: true }
        is_active: { type: boolean }
        price_options:
          type: array
          items: { $ref: "#/components/schemas/PriceOption" }
      required: [id, washing_point_id, name, duration_minutes, is_active, price_options]
    ServiceList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Service" } }
      required: [items]
    ServiceCreate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        description: { type: string, maxLength: 2000 }
        duration_minutes: { type: integer, minimum: 1 }
        picture_url: { type: string, maxLength: 500 }
        price_options:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/PriceOptionInput" }
      required: [name, duration_minutes, price_options]
    ServiceUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        description: { type: string, maxLength: 2000 }
        duration_minutes: { type: integer, minimum: 1 }
        picture_url: { type: string, maxLength: 500 }
        is_active: { type: boolean }
...